requirements.txt
branca==0.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
folium==0.19.4
googlemaps==4.10.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.2
pandas==2.2.3
pdfkit==1.0.0
polyline==2.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
xyzservices==2025.1.0
